[functions][ufo] read notes
one, added function -- specifically the dropdown function -- to the GEN folder for the intention of sharing that dropdown code -- seemed kind of redundant to keep going back to specific pages to pull out that functionality
two, been working on the UFO Dashboard Site -- it's pretty fun making some progress. Working on the mobile after figuring out how the desktop is going to behave. Specifically on how to make the individual nav item at the page bottom into its own link despite sharing the behavior and look of the above link menus